+++ Engine/Companions/ChatGPT/reasoning_effort.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel;
+
+namespace JocysCom.VS.AiCompanion.Engine.Companions.ChatGPT
+{
+	public enum reasoning_effort
+	{
+		[Description("Low")]
+		low = -1,
+		[Description("Medium")]
+		medium = 0,
+		[Description("High")]
+		high = 1,
+	}
+}