+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2025-01-13 v1.16.22
 
 - New: `GenerateQrCodeImage` function. Can be used to send data to a mobile using the camera.
+- New: `ReasoningEffort` option for the `o1` and `o1-mini` reasoning models.
 
 2025-01-12 v1.16.20